Big update!
Added the "Who" page, still WIP.
Optimized CSS styles.
VOptimized the "About" page, since it only had one component and nothing else I integrated (and then removed it as a single) that component with the page.
Fixed a scary bug which was sitting in the ratings management.
Moved the font in the correct folder where it should be.
Minor adjustments.